+++ src/components/DevContainerComponent.ts
@@ -1,7 +1,6 @@
 import { JsonFile, JsonFileOptions } from 'projen';
 import { TypeScriptProject } from 'projen/lib/typescript';
-import { IProjectComponent } from '../types/component';
-import { Scripts } from '../types/script';
+import { IProjectComponent, Scripts } from '../types';
 
 /**
  * Configures the DevContainer templates, settings and npm scripts for the project.
